Yield through the mix walks, and keep the onset strength in 32 bits
mix_build() holds a CPU boost across two walks of the whole library -- every
record scored against every step of the goal, then the database for what is
behind the keys it kept -- and yielded in neither. Scheduling is cooperative,
so that is the UI stopped and the codec not refilling for as long as the walk
takes, which is heard rather than measured. Pass one yields every sector's
worth of records, so the yield lands where the read does; pass two yields per
row. db_summary.c says the same thing about the same kind of pass and is worth
reading beside this.

The onset strength widened to 64 bits for a product that cannot reach 32. A
flux is at most three bands of a magnitude the Goertzel clamps to 46340, so
139020 is the ceiling and 128 times that is a thirteenth of an int -- and the
width bought a software division several times the cost of the one it needed.

+++ apps-ipod/audio/beat_hops.c
@@ -475,9 +475,13 @@ static void beat_run_hop(void)
              * reads 128 and section 4.1's accent -- half again above that
              * peak -- reads 192. Scaling so the peak itself read full would
              * saturate nearly every onset, since an onset only fires by
-             * standing near the peak in the first place. */
+             * standing near the peak in the first place.
+             *
+             * 32 bits is enough: a flux is at most three bands of a
+             * magnitude the Goertzel clamps to 46340, so 139020 is the
+             * ceiling and 128 times that is a thirteenth of an int. */
             int peak = flux_peak[g] > 0 ? (int)flux_peak[g] : 1;
-            int strength = (int)(((int64_t)flux_1[g] * 128) / peak);
+            int strength = (flux_1[g] * 128) / peak;
 
             beat_mark_onset(g, hop_ms_1,
                                 strength > 255 ? 255 : strength);
